Add Rack Attack configuration
